added events for service and characteristics discovery
(serviceDiscovered, characteristicsDiscovered) that enables
caching of gatt information.
The cached objects can be used to initialize a device via addServices() and addCharacteristics(), which return an
array of Noble objects ready for connecting in the usual way.

examples/cache-gatt-discovery.js provides an example for collecting GATT data and persisting the information,
examples/cache-gatt-reconnect.js uses the data to connect to the device using the cached data.

Author: Simon Vogl <simon@voxel.at>
Relate-to: https://github.com/abandonware/noble/pull/29
Change-Id: I68bc5208529f872e5f60d7c0315a6346d24df2fb